requirements.txt
boto==2.48.0
bottlenose==1.1.2
click==6.7
Flask==0.12.2
Flask-Cors==3.0.3
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
lxml==3.8.0
MarkupSafe==1.0
python-amazon-simple-product-api==2.2.11
python-dateutil==2.6.0
six==1.10.0
Werkzeug==0.12.2
